Add `characterEncoding` and `useUnicode` to JDBC properties.
I am proposing the addition of these two properties as part of the example because they required in the DW 1.0 app I am currently working on. It is worth noting that the DB is MySQL, so this is potentially only an issue with MySQL implementations. Omitting `characterEncoding` and `useUnicode` does not cause anything to fail but it prevents the app from properly enforcing UTF-8. A simple example includes persisting and retrieving the ™ symbol.