Jersey Validation Improvements (#1734)
Better Enum Validation Messages

Currently Dropwizard resources return a 404 if a request is made with
an invalid value for a parameter of an Enum type. This change adds
ValidatingEnumParamConverterProvider, which provides converters for all
enums used in method parameters. The converter then provides better
error messages like "query param choice must be one of [OptionA,
OptionB, OptionC]" when invalid values are sent.

Parameter Names in Parse Failure Messages

Currently when an invalid value is passed into a *Param parameter,
Dropwizard returns an error message like "Parameter is not a
number". This change includes AbstractParamConverterProvider along
with a slight enhancement to the *Param classes that returns messages
that include the name of the parameter that failed validation, like
"query param user_id is not a number."

Hibernate ParameterNameProvider

Small change to move most of parameter naming logic from
ConstraintMessage to JerseyParameterNameProvider. This implements the
standard ParameterNameProvider interface that Hibernate uses to
resolve parameter names during validation.

New DurationParam

Include a new DurationParam that can parse String input in the format
supported by Dropwizard's existing Duration class.