Fix request logs with request parameter in layout pattern
Add a safe version of `RequestParameterConverter` which works with async appenders.
It loads request parameters from a cached map rather than trying to load request data
from the original request, which may be closed during the time of a call.  The map is
cached by the  `prepareForDefferedProcessing` method, which invokes before invoking
appenders, so in theory the call to the map should be safe.

As a result, Dropwizard users can use `%reqParameter{something}` in their request
logging patterns.

References #1827, #1672, #1686.