Delegate calculation amount of acceptors and selectors threads to Jetty
The issue is reported in #523. The OS on CI Linux machines returns an
astonishingly big amount of CPUs (~ 128). Jetty changed their algorithm
for calculating the maximum amount of selector and acceptor threads,
see `eclipse/jetty.project@2d52280`. But, in Dropwizard we still set
the amount of acceptors as #CPUs/2 and selectors #CPUs. Looks like
that's too much for Jetty, and it can't handle such a big amount of threads.
Because of that we have random errors on our CI environment and,
what worse, possibly hurt users who actually run their applications on
machines with a big amount of CPU. A solution is to delegate calculating
the amount to Jetty (which has more sane defaults) and document the
defaults.