Run all migration tests in memory
We can run the database in memory with the `close_delay=-1` parameter
instead of the file mode. That parameter means that the database should
be closed only when the JVM exits, not when the JDBC connection is
closed. Hence, the changes performed by one command are visible to the
others in the same JVM, which is exactly what we need.