Support `@UnitOfWork` in sub-resources
Currently, Dropwizard doesn't open transactions in sub-resources as reported
in #1806. The problem is that Dropwizard scans resource methods for
`@UnitOfWork` during resource initilization and sub-resource methods are
not resolved at that time and can't be registered by Dropwizard. The fix is
to defer the lookup of the `@UnitOfWork` annotation on a method until it's
invoked and than cache it. One downside of this approache is that
`ConcurrentMap` doesn't support null values, so we have to use `Optional`
from `UnitOfWork` which can add an additional overhead, but it should be
negligible.