Support disabled and enabled attributes for metrics (#1957)
Metrics 3.2.x allows to configure the set of attributes which can
be disabled for a reporter. It would be nice to provide the ability
to set them via the YAML config the same way we provide that for
metric names.

(cherry picked from commit e4a4fdcda7265a8f58377aad3f120f24038cd97f)