Correctly set up SO_LINGER for the HTTP connector
Jetty's `ServerConnector` accepts `soLingerTime` in milliseconds, rather
than in seconds as `Socket.setSoLinger` does [1]. Dropwizard on the other
hand after parsing the YML configuration and extracting the `soLingerTime` parameter
passes it to Jetty in seconds. As a result, the configuration gets really
confusing for users, because the configuration value is more than the actual
SO_LINGER timeout by 1000 times.

The fix is to pass the configuration value to Jetty's `ServerConnector`
in milliseconds.

[1]: http://www.eclipse.org/jetty/documentation/9.4.x/configuring-connectors.html