Add support for parametrized commands with parametrized configurations
Jackson is able to deserialize parametrized configuration classes, so
we should support configured commands with them. To figure out the type,
we need to check the raw type one of the type arguments and check
if it's can be cast to the `Configuration` class.

Simply put, for the class
`RunDbVacuum extends ConfiguredCommand<ReplicaConfig<Replica>>`
`Generics` should be able to figure out that we need to deserialize
the config as `ReplicaConfig`.