Add tests for fuzzy deserialized enums (#2200)
The class Enums is used in Jackson (FuzzyEnumModule) and in Jersey
(FuzzyEnumParamConverter). Both of these classes have unit tests
and they work around the the Enums class (Jersey requires support of
deserialization from a `fromString` method, Jackson support of
@JsonCreator methods and the @JsonProperty annotation). It would
be to have a unit test which states what kind of deseralizations
it supports.

See #2161 and #1909.