Migrate deprecated classes from commons-lang to commons-text
Several classes and methods from the `commons-lang` package
like `StrSubstitutor` and `StringUtils.getLevenshteinDistance` have
been moved to a separate library `commons-text`. We should follow
this deprecation warning and move our dependencies to the new library.

Unfortunately, it's not possible to make this change for the 1.2.1
version, because it changes the public API.