jersey-client: Add possibility to mark an HTTP entity as repeatable
Currently, `DropwizardApacheConnector` when configured with the
chuncked encoding always assumes that the source entity from the
Jersey client is not repeatable. That's true in the default case
when we don't know from where the source of the input. Unfortunately,
the Apache HTTP client can't retry unrepeatable entities, which poses
challenges in environments where retries are common (e.g. microservices).

The only alternative is to disable the chunked encoding which is not
always desirable, because it requires additional buffering of the
entity in memory.

The solution of this issue to add a parameter which allows to configure
the connector to use chunked encoding, but mark them as repeatable.
In this case the caller is responsible to make sure that entites are
coming from a retriable source (e.g memory array, file on disk).