Don't shut down asynchronous executor in Jersey (#2221)
If a user customized a Jersey client, Jersey creates a new copy
of the instance. Jersey has a mechanism for detecting leaks, and Jetty
will invoke the `dispose` method on the old client. Unfortunately,
during a copy the executor is shared. Therefore, the user will not be
able to perform async requests anymore.

The executor is managed by the Dropwizard environment, so we can leave
with the fact that it's closed by a kill signal.

Fixes #2218