Add a TCP and UDP log appenders to Dropwizard
Add support for sending log events to a log management server via TCP or
UDP. This is useful for applications that want to ship logs to a remote
machine. We recently have added support for formatting messages as JSON
and in combination with TCP/UDP logging it can be an interesting solution
for a direct log shipping from Dropwizard without 3rd-party modules.

From the implementation side TcpSocketAppenderFactory and
UdpSocketAppenderFactory are discoverable and work with generic logging
events. That means they can be configured in the config file both for
general and request logging.

logging:
  level: INFO
  appenders:
    - type: tcp
      host: localhost
      port: 24562

server:
  requestLog:
    appenders:
      - type: tcp
        host: localhost
        port: 24562

From the reliability side, the appenders extend
ResilientOutputStreamBase. That means in case if the connections drops
because of network glitch or a DNS failure, it will back off and try to
reopen a new connection automatically.