Make UnitOfWork context publically accessible
The UnitOfWork framework is designed to allow multiple Hibernate bundles to be configured. The bundle for a particular unit is determined at runtime but is not made available to other components. This information is particularly useful to classes such as io.dropwizard.hibernate.AbstractDAO which use the current session to execute queries.

This set of changes introduces a static thread container to make the context determined inside of UnitOfWorkAspect available globally.