Strip the Content-Length header after decompressing HTTP requests
`BiDiGzipHandler` decompresses HTTP requests with compressed data, so
for handlers down the stack the content looks like a plain-text data
and they don't need to bother with decompressing in the application
level logic. Unfortunately, the handler doesn't strip the original
`Content-Length` header from requests, and requests return the content
length of compressed content instead of uncompressed. This can create
confusing situations for applications which check the content length
of requests for future processing. A classic example, a reverse proxy
handler which parses a request, enriches it and sends it to an origin
server. In this case, the proxy handler will send plain data to the
origin server, but with a misleading `Content-Length` header and the
origin server will reject the request as malformed.

A solution is to strip the `Content-Length` header from the request
and override the `getContentLength` method of `HttpServletRequest` to
specify an unknown length. In this cases the proxy handler will pass
the request as chunked to origin server (albeit without the
`Transfer-Encoding` header), and it will be correctly handled. This
seems to be a correct behaviour, because we don't know the length of
decoded content beforehand.

(cherry picked from commit 6d86f8a)