Upgrade Jackson to 2.8.11 in 1.1.* to address a CVE (#2270)
https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2017-17485

According to the CVE, maliciously crafted JSON input can allow a
remote code execution, if it's passed directly to the `readValue`
method of `ÒbjectMapper`. The blacklist of deserialized types is
ignored if the Spring libraries are available in the classpath.

Dropwizard doesn't use Spring, but some end users use Spring along
with Dropwizard, so we should give them a simple way to protect
their applications.