Make sure test Jersey test client uses Dropwizard's ObjectMapper
Currently the Jersey client produced by `DropwizardAppRule` uses
a simple Jackson's ObjectMapper. This could be an issue when
a serialized/deserialized entity requires additional modules (Java 8
`Optional`, `DateTime`, `ParamaterNames` module etc...), which are
added automatically in Dropwizard's `ÒbjectMapper` and used for the
server side. A fix is too prodouce a client with a Jackson binder
which uses the `ObjectMapper` instance from the test environment.