Correctly handle the default context path when logging endpoints
The default context path is not an empty string, but slash. Currently,
a slash is considered as a custom context path and printed out along with
the endpoint with a slash. So, for a typical endpoint we output two
slashes, which is not great.

A fix is to consider a slash as a default context path and don't print
it out.

This is a follow-up for #2254