Reuse common code across the appenders
Add a new class `AsbtractOutputStreamAppenderFactory` which allows
us to build appenders based on `ÒutputStreamAppender`, e.g. console,
file, tcp, udp. It adds a new abstract method for building an appender
which is implemented by actual factories and takes care of enchancing
this appender with the layout encoder, filter factories and an async
appender.