Allow lazy initialization of resources in ResourceTestRule
Instead of eagerly initializing resource instances added to `ResourceTestRule` or
`Resource`, we now allow using a `Supplier<T>` which can be used to lazily initialize
the resources or providers used in `ResourceTestRule`.

Additionally, this change set adds an adapter for `MockitoJUnit` which allows to use
a Mockito JUnit Rule in `RuleChain` for defining an execution order for rules.

Closes #1800