Remove expected exceptions from tests
Replace them with AssertJ's `assertThrown`. Modern testing libraries
seems to favour this approach, because it provides flexibility in
verifying the cause of the error. For example, JUnit5 doesn't support
it anymore the `expected` attribute in the `Test` annotation and the
`ExpectedException` rule. If we ever switch to JUnit5 for testing,
migration would be easier.