Add TLS socket logging appender
Add support for sending logs to a server that requires a secure means
of communication via SSL/TLS. This change adds a new logging factory
which produces a new socket factory for creating secure client SSL/TLS
sockets.

* The default configuration allows to communicate with a server with
a certificate signed by a CA. In that case the administrator only need
to change the port and the type of the appender.

* If the server is deployed in a internal network with a self-signed
certificate, the factory allows to specify a path to a trust store
with the server certificate.

* If the server requires the client to authenticate itself, the factory
allows to specify a path to a key store with a client certificate
that will be sent to the server during negotiation.

* The factory allows to configure allowed cipher suites and SSL/TLS
protocols.