httpclient: Try to load keystore with the specified JCE provider
If the user configured the HTTP client to use a JCE provider, we
should try to load the keystore with it, so the users can load
handle complex keystore formats not support by the standard JCE
providers or to take advantage of better performance characteristics
of the custom provider.

In case the new provider doesn't support the keystore format, we
fallback to the default discovery algorithm.