Fix ResourceURL.isDirectory to handle Jar paths that are URL-escaped.
The heuristic inside the "jar" codepath in ResourceURL.isDirectory is
splitting the filename from the URL by String indexing.  This means that
if the file name would be URL-escaped (e.g. if it has a space in it) then
relativeFilePath is still escaped.  This manifests as zipEntry being null,
which then causes NullPointerException -> IOException ->
ResourceNotFoundException.

Fix this by using entry.getName() instead of taking a URL substring.
This works because jarConnection has been formed by correctly parsing the URL.