Change fuzzy enum deserializer to be cacheable.
During load tests, there is significant lock contention in Jackson when
using the FuzzyEnumModule. For each serialized enum type, there is a
synchronized block that attempts to create a serializer in the cache and
ultimately since the PermissiveEnumDeserializer isn't marked as
cacheable, the underlying deserializer is recreated each time. Update
the PermissiveEnumDeserializer to be cacheable (similar to
EnumDeserializer from Jackson) to remove significant overhead during
deserialization.