ConnectorProvider Not Set Silently
When the JerseyClientBuilder receives a null value for
ConnectorProvider, it creates a value and then sets that field.
Subsequent builds then use the same ConnectorProvider. This change
makes it so a null value will not be automatically set, instead users
need to explicitly set ConnectorProvider to use the same one accross
builds.