Catch EofException at the jetty handler level
This change is meant to address https://github.com/dropwizard/dropwizard/issues/2482

Because the EofException is thrown when the handler is trying to write the response, an exception mapper can't solve this issue. In the future, for a more robust solution, we should allow logging levels to be configurable here.