Nested calls to @UnitOfWork methods cause inconsistent behaviour (#2913)
When a method annotated with @UnitOfWork calls into another method that
is annotated like that, the transaction is prematurely committed when
the inner method returns.

This leads to difficult to debug situations and should be prevented.
Either an exception has to be raised as soon as possible or nesting
needs to do the sane thing.

The implementation checks if a session is already associated with this
thread. If that is not the case (non-nested), the original code paths
are invoked. If there exists already a session, its state is validated,
if the session is compatible with the requested state, the unit of work
is joined, if not an IllegalStateException is raised.