[views] Ability to override locale detection
In our application we want to have a custom application specific locale
detection procedure based on query parameters of the HTTP request.
Currently, the ViewMessageBodyWriter.detectLocale method is package
private which makes it quite cumbersome to override.

Make ViewMessageBodyWriter.detectLocale protected

We will be able to register our own ViewMessageBodyWriter implementation
which overrides the `detectLocale` method in a nice way.