Add config settings for tasks and health check resources
This change set adds configuration settings for the tasks resource and the health check
resources in the Dropwizard admin interface.

The `printStackTraceOnError` (default: false) setting allows to control whether to expose
the full exception stack trace of a failing task.

The `minThreads`, `maxThreads`, and `workQueueSize` configuration settings allow to control
the thread settings for health checks.

Closes #3036