Fix regression in OptionalInt/Long/Double handling.
DropWizard 2.x changed the behavior of OptionalInt/Long/Double
parameters in two ways. First, empty values used to result in the
@DefaultValue of a parameter being used. In DropWizard 2.x, this changed
to return OptionalXYZ.empty() instead.

Additionally, using an invalid value for a @DefaultValue parameter used
to fail fast when the Jersey resource was added. This changed with more
lax parsing in DropWizard 2.x.

Update to preserve the previous behavior of empty string parameter
values and invalid @DefaultValue values. This should avoid issues with
adoption of DropWizard 2.x.

Fixes #3123.