Executor metrics shouldn't contain fmt string.
Update ExecutorServiceBuilder to remove the format string from the
metric name. Currently, a thread pool named 'jersey-client-foo-%d' will
have metrics 'jersey-client-foo-%d.created',
'jersey-client-foo-%d.terminated', and 'jersey-client-foo-%d.running'
added to the metrics registry. The '-%d' format string is only meant to
ensure that threads are uniquely named and shouldn't be included in the
name of the metric.

Update ExecutorServiceBuilder to strip the format portion of the
nameFormat from the metric name.

Fixes #3139.