Fix regression in `AbstractParam` handling
In Dropwizard 1.3.x and earlier, Jersey interpreted empty strings in parameters (`@CookieParam`, `@QueryParam`, etc.) identical to `null` as missing and used the default value (`@DefaultValue`) as fall-back if it existed.

This change set restores the described behavior in Dropwizard 2.x for the classes inheriting from `AbstractParam`.

Keep in mind that `AbstractParam` and its child classes are deprecated and will be removed in a later Dropwizard version in favor of using `Optional<T>` directly.

Refs #3123
Refs #3134