Enable maven-dependency-plugin analysis.
Update dropwizard to run maven-dependency-plugin analyze-only goal as
part of the build. This will automate detection of new dependencies as
well as when dependencies are no longer in use. In some cases there may
be some false positives (seen primarily when only using an import from a
dependency), however overall this should improve the quality of the POM
files for dropwizard.