Support comments in DiscoverableSubtypeResolver (#2992)
Add the ability to have comments in the `DiscoverableSubtypeResolver` class resources file.

# Problem

Currently if you have any other lines in the discoverable resource files you get a bunch of log lines detailing "cannot load this class".  Example:
```
# this is the foo class
org.github.Foo
```

# Solution

Add the ability to allow for comments in the resource file.  Comments are lines that start with `#`.

# Result

No more log lines for lines that begin with `#` detailing that it's not a class when parsing the class resource files.