Allow overriding default media type in AssetServlet/AssetBundle (#3269)
Sometimes the default media type of `text/html` in the `AssetServlet` and `AssetBundle` might not be the best fit for every use case.

This PR exposes a setting to specify the default media type for `AssetBundle` and `AssetServlet` and override the default of `text/html`.

(cherry picked from commit 45a26e39bd8c3d382d43e48963861cae077ac519)