Reduce flakiness of LayoutIntegrationTests
Two of the tests were waiting for log output using a one-off call to
Thread.sleep(). In addition to being brittle, the tests would fail with
an NPE (while trying to assert on the "timestamp" field) that doesn't
point to the cause of the failure directly.

Also removed the redundant assertThat(jsonNode).isNotNull() as
objectMapper.readTree() will never return null in Jackson 2.10.x, see
https://github.com/FasterXML/jackson-databind/issues/2211