Add ConfigOverride for random application ports during tests (#3561)
Problem:

We are doing a lot of integration testing where we need to start the application. We primarily use the config overrides instead of a `test-config.yaml` to do the configuration. Our config files of the test module often only contain the configuration for the random ports:
```yaml
server:
  applicationConnectors:
    - type: http
      port: 0
  adminConnectors:
    - type: http
      port: 0
```

It would be nice to be able to use random ports with the `DropwizardAppRule`/`DropwizardAppExtension` programmatically.

Solution:

Add a Config Override that configures the default application and admin connectors to use port `0`.

```java
private static final DropwizardAppExtension<MyConfiguration> EXTENSION =
    new DropwizardAppExtension<>(MyApplication.class,
        // no config file
        null,
        // adds all necessary config overrides for random ports
        randomPorts());
```

Closes #3561