Add method to `AbstractDAO` to get `NamedQuery` in a type-safe manner (#3978)
`AbstractDAO` offers a way to easily get named queries but does so in a way that requires casting the returned `Query` to the correct type.

Added a method to `AbstractDAO` which returns a named query in a type-safe way, `namedTypedQuery(String)`. This methods uses `Session#createNamedQuery(String, Class)` internally where the generic type is retrieved via `AbstractDAO#getEntityClass()`.