Use Liquibase Scope for providing objects to CustomChanges (#4201)
###### Problem:
When applying big data migrations with a Liquibase `CustomChange`, one cannot use Hibernate or JDBI at the moment.

###### Solution:
As of Liquibase 4.1.0, the `Scope` class for global configuration got an `enter` method, that takes a `Map<String, Object>` of objects, that are then available through the `get` method of the `Scope` class. Therefore the `MigrationsBundle` can be extended with a method for providing objects, that should be available inside the `CustomChange`.

###### Result:
When inside the `CustomChange`, the new `Scope` is entered and the provided objects are available.