Prepare ByteStreams for deprecation in Java 9+
In Java 9+ we can use `InputStream#readAllBytes()` to slurp an `InputStream`
into a `String`. Prepare for this future refactor by using
`ByteStreams#toByteArray()` to convert `InputStream`s to a byte array whenever
we're converting an `InputStream` to a `String`.

`ByteStreams#copy()` is only used by `dropwizard-util`, so deprecate it and
replace it with a package-private method.