Stop reporting `accept()` failures during `TcpServer` shutdown
When `TcpServer#tearDown()` is called, it marks its acceptor thread interrupted
and closes the `ServerSocket`.

This causes the `accept()` call to throw a `SocketException` which triggers a
call to `fail()`. This doesn't actually fail the test because it's not called
in the main thread.

Silence this log noise by first checking to see if the current thread has been
interrupted (i.e. we are shutting down and an exception is expected).