Remove duplicate logfile from dropwizard-logging tests
This was added in cc53136ae8754c20df268b6ea1a0cf3528618450 but isn't
used by the tests.

It causes errors to be logged during tests as the `currentLogFilename`
is identical to another log configuration in the same file.

`FileAppenderFactoryTest` tests the behaviour of `FileAppenderFactory`
directly without roundtripping through the YAML configuration parser.

Remove the duplicate log file definition.