Remove inconsistency in CsvReporterFactory's directory handling
`CsvReporterFactory#build(MetricRegistry)` does all its validation
directory on the `file` field but uses the result of `getFile()` when
actually constructing the `CsvReporter`.

If someone were to subclass and override `getFile()` this would lead to
confusing behaviour.

Use `getFile()` instead of directly accessing the field.

In addition, the `file` variable in `build(MetricRegistry)` shadowed the
`file` field. Rename it to `directory` to avoid this.