Simplify switch statement in `ResourceURL#getLastModified()`
This is a bit of a noop change - I don't think there were any gaps in
the control flow.

Sonar was struggling to work out whether cases could fall-through, but
they can't.

Still, moving the `return 0` to the end of each case statement makes it
very clear what the behaviour will be.