Make `@UnitOfWork` a repeatable annotation (#4198)
###### Problem:
In some situations, there is the need to access multiple databases in a Jersey resource method.
At the moment, only one `@UnitOfWork` annotation can be provided to a Jersey method and therefore only one database can be managed automatically.

###### Solution:
Starting from Java 8, annotations can be repeatable. So make `@UnitOfWork` a repeatable annotation.
If multiple annotations with the same value are provided, the last one gets used.

###### Result:
When a Jersey resource method is annotated with multiple `@UnitOfWork` annotations, for each of these a `UnitOfWorkAspect` gets instantiated and the hibernate `Session`s all get managed automatically.