Hibernate: DualSessionFactory (#3580)
# Problem:
Modern RDBMS systems often provide read-replica clusters with separate access points. With regards to Hibernate, this requires two SessionFactory instances.
One factory is for the primary/writer data source and a separate factory is necessary for the read-replica data source. 

This requires the usage of the UnitOfWork name property whenever the read-replica is preferred. Ideally, no adjustments to the UnitOfWork properties should be necessary if the readOnly is properly set across endpoints.

# Solution:
This commit provides a SessionFactory implementation that contains both a writer/primary factory and a reader factory. 

The DualSessionFactory comes with separate HibernateBundle, UnitOfWorkApplicationListener, and UnitOfWorkAspect classes.

Prior to usage, the UnitOfWorkAspect class prepares the DualSessionFactory based on the UnitOfWork annotation readOnly property. 
When the dual HibernateBundle is dropped into an application, the UnitOfWork requires no changes across endpoints assuming that the readOnly property is already set accordingly.

# Result:
This commit contains new DualSessionFactory, HibernateBundle, UnitOfWorkApplicationListener, and UnitOfWorkAspect classes.