Use Grizzly connector to avoid using HttpUrlConnector workaround (#5403)
Refs #4404 

# Problem:
The `JerseyClientBuilder` provided in `DropwizardAppExtension` uses the default connector provider, i.e. `HttpUrlConnectorProvider`. As stated in [this comment](https://github.com/eclipse-ee4j/jersey/issues/4825#issuecomment-925836004), this connector provider will not support the HTTP PATCH method on Java 16+.

# Solution:
Use an other connector provider, which natively supports the HTTP PATCH method. This PR uses the `GrizzlyConnectorProvider` because it passes most of our tests.

# Breaking change:
This PR technically is a breaking change, since the obtained client builder is configured differently. But as the HTTP PATCH method will not work on Java 16+, this is a tradeoff IMO. I think we could introduce this change because a different `JerseyClientBuilder` can be obtained easily and hence we should have a client working on Java 16+ as a default.