JavaDoc enhancements (#5749)
* Fix JavaDoc warnings
* Fix JavaDoc warnings in `dropwizard-jackson`
* Fix JavaDoc warnings in `dropwizard-validation`
* Fix JavaDoc warnings in `dropwizard-configuration`

This PR is meant as a starting point to fix JavaDoc errors during the run of the compiler and during javadoc generation via the `maven-javadoc-plugin`.

@dropwizard/committers I think up-to-date JavaDoc is always a good thing, although this might take some time to accomplish. I've now corrected the errors in the `dropwizard-util` module. Do you think it's worth doing so for all other modules?

I've experienced, that the invocation of the `javadoc:jar` target on a JDK 17 causes errors, even with our `<doclint>none</doclint>` settings. The errors vanish when passing `-Ddoclint=none` on the command line. Maybe I'm missing something here, but we should be able to build the project on a JDK 17.