Speed up LogbackClassicRequestLogFactoryTest (#8014)
In the test class LogbackClassicRequestLogFactoryTest, there are three tests but all of them are not modifying requestLog or other related classes and they are just accessing requestLog. Repeatly creating the requestLog object from the yaml file will increase the test runtime.

The solution is to just create the requestLog once before running all tests in the test class LogbackClassicRequestLogFactoryTest.

There will be no repeated cost when creating the requestLog by justing creating it once. The test runtime can also be improved.

Co-authored-by: Jochen Schalanda <jochen@schalanda.name>